+++ modules/sdk-core/src/bitgo/trading/network/index.ts
@@ -1,2 +1,4 @@
+export * from './encrypt';
+export * from './decrypt';
 export * from './network';
 export * from './types';