+++ modules/sdk-lib-mpc/src/tss/ecdsa-dkls/dkg.ts
@@ -1,27 +1,76 @@
 import { KeygenSession, Keyshare, Message } from '@silencelaboratories/dkls-wasm-ll-node';
-import { DeserializedBroadcastMessage, DeserializedMessages, DkgState } from './types';
-import { decode } from 'cbor';
+import { DeserializedBroadcastMessage, DeserializedMessages, DkgState, RetrofitData } from './types';
+import { decode, encode } from 'cbor-x';
+import { bigIntToBufferBE } from '../../util';
+import { Secp256k1Curve } from '../../curves';
 
 export class Dkg {
   protected dkgSession: KeygenSession | undefined;
   protected dkgSessionBytes: Uint8Array;
   protected dkgKeyShare: Keyshare;
+  protected keyShareBuff: Buffer;
   protected n: number;
   protected t: number;
   protected chainCodeCommitment: Uint8Array | undefined;
   protected partyIdx: number;
   protected dkgState: DkgState = DkgState.Uninitialized;
+  protected dklsKeyShareRetrofitObject: Keyshare | undefined;
+  protected retrofitData: RetrofitData | undefined;
 
-  constructor(n: number, t: number, partyIdx: number) {
+  constructor(n: number, t: number, partyIdx: number, retrofitData?: RetrofitData) {
     this.n = n;
     this.t = t;
     this.partyIdx = partyIdx;
     this.chainCodeCommitment = undefined;
+    this.retrofitData = retrofitData;
   }
 
   private _restoreSession() {
     if (!this.dkgSession) {
       this.dkgSession = KeygenSession.fromBytes(this.dkgSessionBytes);
+    }
+  }
+
+  private _createDKLsRetrofitKeyShare() {
+    if (this.retrofitData) {
+      if (!this.retrofitData.xShare.y || !this.retrofitData.xShare.chaincode || !this.retrofitData.xShare.x) {
+        throw Error('xShare must have a public key, private share value, and a chaincode.');
+      }
+      if (this.retrofitData.bigSiList.length !== this.n - 1) {
+        throw Error("bigSiList should contain the other parties' Si's");
+      }
+      const bigSList: Array<Array<number>> = [];
+      const xiList: Array<Array<number>> = [];
+      let j = 0;
+      for (let i = 0; i < this.n; i++) {
+        if (i === this.partyIdx) {
+          const secp256k1 = new Secp256k1Curve();
+          bigSList.push(
+            Array.from(bigIntToBufferBE(secp256k1.basePointMult(BigInt('0x' + this.retrofitData.xShare.x))))
+          );
+        } else {
+          bigSList.push(Array.from(Buffer.from(this.retrofitData.bigSiList[j], 'hex')));
+          j++;
+        }
+        xiList.push(Array.from(bigIntToBufferBE(BigInt(i + 1), 32)));
+      }
+      const dklsKeyShare = {
+        total_parties: this.n,
+        threshold: this.t,
+        rank_list: new Array(this.n).fill(0),
+        party_id: this.partyIdx,
+        public_key: Array.from(Buffer.from(this.retrofitData.xShare.y, 'hex')),
+        root_chain_code: Array.from(Buffer.from(this.retrofitData.xShare.chaincode, 'hex')),
+        final_session_id: Array(32).fill(0),
+        seed_ot_receivers: new Array(this.n - 1).fill(Array(32832).fill(0)),
+        seed_ot_senders: new Array(this.n - 1).fill(Array(32768).fill(0)),
+        sent_seed_list: [Array(32).fill(0)],
+        rec_seed_list: [Array(32).fill(0)],
+        s_i: Array.from(Buffer.from(this.retrofitData.xShare.x, 'hex')),
+        big_s_list: bigSList,
+        x_i_list: this.retrofitData.xiList ? this.retrofitData.xiList : xiList,
+      };
+      this.dklsKeyShareRetrofitObject = Keyshare.fromBytes(encode(dklsKeyShare));
     }
   }
 
@@ -61,9 +110,14 @@ export class Dkg {
     }
     if (typeof window !== 'undefined') {
       const initDkls = require('@silencelaboratories/dkls-wasm-ll-web');
-      await initDkls();
+      await initDkls.default();
     }
-    this.dkgSession = new KeygenSession(this.n, this.t, this.partyIdx);
+    this._createDKLsRetrofitKeyShare();
+    if (this.dklsKeyShareRetrofitObject) {
+      this.dkgSession = KeygenSession.initKeyRotation(this.dklsKeyShareRetrofitObject);
+    } else {
+      this.dkgSession = new KeygenSession(this.n, this.t, this.partyIdx);
+    }
     try {
       const payload = this.dkgSession.createFirstMessage().payload;
       this._deserializeState();
@@ -77,9 +131,10 @@ export class Dkg {
   }
 
   getKeyShare(): Buffer {
-    const keyShareBuff = Buffer.from(this.dkgKeyShare.toBytes());
-    this.dkgKeyShare.free();
-    return keyShareBuff;
+    if (!this.keyShareBuff) {
+      throw Error('Can not get key share, DKG is not complete yet.');
+    }
+    return this.keyShareBuff;
   }
 
   handleIncomingMessages(messagesForIthRound: DeserializedMessages): DeserializedMessages {
@@ -117,6 +172,11 @@ export class Dkg {
       }
       if (this.dkgState === DkgState.Round4) {
         this.dkgKeyShare = this.dkgSession.keyshare();
+        if (this.dklsKeyShareRetrofitObject) {
+          this.dkgKeyShare.finishKeyRotation(this.dklsKeyShareRetrofitObject);
+        }
+        this.keyShareBuff = Buffer.from(this.dkgKeyShare.toBytes());
+        this.dkgKeyShare.free();
         this.dkgState = DkgState.Complete;
         return { broadcastMessages: [], p2pMessages: [] };
       } else {