+++ modules/sdk-lib-mpc/src/tss/ecdsa-dkls/util.ts
@@ -2,7 +2,7 @@ import { Signature } from '@noble/secp256k1';
 import { HDTree, Secp256k1Bip32HdTree, Secp256k1Curve } from '../../curves';
 import { bigIntFromBufferBE, bigIntToBufferBE } from '../../util';
 import { DeserializedDklsSignature } from './types';
-import { decode } from 'cbor';
+import { decode } from 'cbor-x';
 import * as secp256k1 from 'secp256k1';
 import { Hash, createHash } from 'crypto';
 
@@ -27,8 +27,8 @@ export function combinePartialSignatures(round4MessagePayloads: Uint8Array[], rH
   const sig = new Signature(r, s);
   const normalizedSig = sig.normalizeS();
   return {
-    R: new Uint8Array(bigIntToBufferBE(normalizedSig.r)),
-    S: new Uint8Array(bigIntToBufferBE(normalizedSig.s)),
+    R: new Uint8Array(bigIntToBufferBE(normalizedSig.r, 32)),
+    S: new Uint8Array(bigIntToBufferBE(normalizedSig.s, 32)),
   };
 }