+++ modules/abstract-eth/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [21.1.1](https://github.com/BitGo/BitGoJS/compare/@bitgo/abstract-eth@21.1.0...@bitgo/abstract-eth@21.1.1) (2024-05-01)
+
+**Note:** Version bump only for package @bitgo/abstract-eth
+
+# [21.1.0](https://github.com/BitGo/BitGoJS/compare/@bitgo/abstract-eth@21.0.0...@bitgo/abstract-eth@21.1.0) (2024-04-25)
+
+### Features
+
+- **abstract-eth:** use non packed encode for V4 wallet ([54fd3a8](https://github.com/BitGo/BitGoJS/commit/54fd3a8dbafb14bdeae78204c122885d288d48b6))
+
 # [21.0.0](https://github.com/BitGo/BitGoJS/compare/@bitgo/abstract-eth@20.1.6...@bitgo/abstract-eth@21.0.0) (2024-04-24)
 
 ### Bug Fixes