+++ modules/sdk-coin-algo/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.8](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-algo@2.1.7...@bitgo/sdk-coin-algo@2.1.8) (2024-05-01)
+
+**Note:** Version bump only for package @bitgo/sdk-coin-algo
+
+## [2.1.7](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-algo@2.1.6...@bitgo/sdk-coin-algo@2.1.7) (2024-04-25)
+
+**Note:** Version bump only for package @bitgo/sdk-coin-algo
+
 ## [2.1.6](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-algo@2.1.5...@bitgo/sdk-coin-algo@2.1.6) (2024-04-24)
 
 ### Bug Fixes