+++ modules/sdk-coin-atom/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [12.0.10](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-atom@12.0.9...@bitgo/sdk-coin-atom@12.0.10) (2024-05-01)
+
+**Note:** Version bump only for package @bitgo/sdk-coin-atom
+
+## [12.0.9](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-atom@12.0.8...@bitgo/sdk-coin-atom@12.0.9) (2024-04-25)
+
+**Note:** Version bump only for package @bitgo/sdk-coin-atom
+
 ## [12.0.8](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-atom@12.0.7...@bitgo/sdk-coin-atom@12.0.8) (2024-04-24)
 
 **Note:** Version bump only for package @bitgo/sdk-coin-atom