+++ modules/sdk-coin-eth/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [24.1.1](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-eth@24.1.0...@bitgo/sdk-coin-eth@24.1.1) (2024-05-01)
+
+### Bug Fixes
+
+- add txHex for eth CCR tx ([780929a](https://github.com/BitGo/BitGoJS/commit/780929a1df6e0a958102fa7b93d7056e593dbc52))
+
+# [24.1.0](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-eth@24.0.0...@bitgo/sdk-coin-eth@24.1.0) (2024-04-25)
+
+### Features
+
+- **abstract-eth:** use non packed encode for V4 wallet ([54fd3a8](https://github.com/BitGo/BitGoJS/commit/54fd3a8dbafb14bdeae78204c122885d288d48b6))
+
 # [24.0.0](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-eth@23.1.6...@bitgo/sdk-coin-eth@24.0.0) (2024-04-24)
 
 ### Bug Fixes