+++ modules/utxo-bin/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.47.12](https://github.com/BitGo/BitGoJS/compare/@bitgo/utxo-bin@2.47.11...@bitgo/utxo-bin@2.47.12) (2024-05-01)
+
+**Note:** Version bump only for package @bitgo/utxo-bin
+
+## [2.47.11](https://github.com/BitGo/BitGoJS/compare/@bitgo/utxo-bin@2.47.10...@bitgo/utxo-bin@2.47.11) (2024-04-25)
+
+**Note:** Version bump only for package @bitgo/utxo-bin
+
 ## [2.47.10](https://github.com/BitGo/BitGoJS/compare/@bitgo/utxo-bin@2.47.9...@bitgo/utxo-bin@2.47.10) (2024-04-24)
 
 **Note:** Version bump only for package @bitgo/utxo-bin